# specs/001-rag-chatbot/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: RAG Chatbot API for Physical AI Book
  version: 1.0.0
  description: API for indexing book content, general chat, selected text chat, and translation for the Physical AI Book, built with Next.js and integrated with Supabase and Gemini.
servers:
  - url: /api
    description: Base URL for the Next.js RAG Chatbot API
tags:
  - name: Indexing
    description: Operations related to document indexing
  - name: Chat
  - name: Authentication
    description: Operations related to user authentication via Supabase
  - name: Translation
    description: Operations for text translation

paths:
  /index:
    post:
      summary: Trigger document indexing
      operationId: triggerIndexing
      tags:
        - Indexing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                force_reindex:
                  type: boolean
                  description: Whether to force a full re-index of all documents.
                  default: false
      responses:
        '200':
          description: Indexing process initiated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Document indexing initiated.
        '500':
          description: Internal server error during indexing
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string

  /chat:
    post:
      summary: Engage in general chat with the RAG chatbot
      operationId: generalChat
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                session_id:
                  type: string
                  format: uuid
                  description: Optional UUID for continuing a chat session.
                query:
                  type: string
                  description: The user's chat query.
                target_language:
                  type: string
                  enum: [ "en", "ur" ]
                  default: "en"
                  description: Optional target language for the chatbot's response (e.g., "ur" for Urdu).
      responses:
        '200':
          description: Chat response with citations and optional translation
          content:
            application/json:
              schema:
                type: object
                properties:
                  response_en:
                    type: string
                    example: The capital of France is Paris, according to Document A [1].
                    description: English response from the chatbot.
                  response_translated:
                    type: string
                    example: فرانس کا دارالحکومت پیرس ہے، دستاویز A [1] کے مطابق۔
                    description: Translated response (if target_language is provided).
                  citations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Citation'
        '400':
          description: Invalid request
        '500':
          description: Internal server error

  /chat/selected:
    post:
      summary: Ask a question based on selected text
      operationId: selectedTextChat
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
                - selected_text
              properties:
                session_id:
                  type: string
                  format: uuid
                  description: Optional UUID for continuing a chat session.
                query:
                  type: string
                  description: The user's question related to the selected text.
                selected_text:
                  type: string
                  description: The text selected by the user from the document.
                target_language:
                  type: string
                  enum: [ "en", "ur" ]
                  default: "en"
                  description: Optional target language for the chatbot's response (e.g., "ur" for Urdu).
      responses:
        '200':
          description: Chat response based on selected text with optional translation
          content:
            application/json:
              schema:
                type: object
                properties:
                  response_en:
                    type: string
                    example: Based on the selected text, the answer is...
                    description: English response from the chatbot.
                  response_translated:
                    type: string
                    example: منتخب متن کی بنیاد پر، جواب یہ ہے...
                    description: Translated response (if target_language is provided).
        '400':
          description: Invalid request
        '500':
          description: Internal server error

  /auth/register:
    post:
      summary: Register a new user via Supabase Auth
      operationId: registerUser
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAuth'
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          description: User with email already exists
        '422':
          description: Validation error

  /auth/login:
    post:
      summary: Authenticate and log in a user via Supabase Auth
      operationId: loginUser
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAuth'
      responses:
        '200':
          description: User logged in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Invalid credentials

  /translate:
    post:
      summary: Translate a given text using Gemini
      operationId: translateText
      tags:
        - Translation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
                - target_language
              properties:
                text:
                  type: string
                  description: The text to be translated.
                target_language:
                  type: string
                  enum: [ "en", "ur" ]
                  description: The language to translate the text into.
      responses:
        '200':
          description: Translated text
          content:
            application/json:
              schema:
                type: object
                properties:
                  translated_text:
                    type: string
                    example: یہ ایک ترجمہ شدہ متن ہے۔
        '400':
          description: Invalid request
        '500':
          description: Internal server error

components:
  schemas:
    Citation:
      type: object
      properties:
        source:
          type: string
          description: Title or identifier of the source document.
          example: "Introduction to ROS"
        context:
          type: string
          description: Relevant snippet of text from the source.
          example: "ROS provides libraries and tools to help software developers create robot applications."
        url:
          type: string
          format: uri
          description: Optional URL to the specific source location.
          example: "docusaurus/docs/module1-ros/intro-to-ros.md"
    UserAuth:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 6 # Supabase default minimum password length
    UserResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
        email:
          type: string
          format: email
          example: user@example.com
        message:
          type: string
          example: User authenticated successfully.